refactor(inference): extract Metal GQA layer encoder
Fixes #841
