+++ crates/inference/src/forward/metal_qwen35/inner/layers/mod.rs
@@ -1,4 +1,5 @@
 mod gdn;
+mod gqa;
 
 use super::*;
